Extended proto file to include Reconnection Message type
In the proto file, the message type can now be set to a reconnect message.

These reconnect messages are enqueued by the tcp_connection when a reconnect happens (not sent through the TCPConnection, only for internal use). This can then be registered by the controller and handled accordingly.

+ Minor cleanup relating to the previous commit.

Added comment to the ReconnectionType enum in the proto file.