tcp_io_device MetaData: In setData, replace &(*it) with a separate variable for the value of *it. Make the arg const ref.
The gcc compiler doesn't allow taking the address of the value in the iterator. "Error taking address of rvalue". Therefore, we copy the value to a local variable and take its address. This also allows us to make the method argument const ref (so that the entire vector is not copied). Even the Visual Studio compiler was complaining about taking the address of a const value.